+++ inst/templates/dispatch-c/tools/kernels/kernel_avx512.c
@@ -6,7 +6,7 @@
 #include <simde/x86/avx512.h>
 
 #include "kernel_common.h"
-#include "simd_dispatch.h"
+#include "kernel_api.h"
 
 size_t rsd_count_nonzero_avx512(const uint8_t *x, size_t n) {
     size_t i = 0;
@@ -55,7 +55,22 @@ void rsd_convolve1d_avx512(const double *a, size_t na, const double *b, size_t n
     }
 }
 
+static void rsd_count_nonzero_avx512_invoke(void *call) {
+    RsdCountNonzeroCall *args = (RsdCountNonzeroCall *)call;
+    args->result = rsd_count_nonzero_avx512(args->x, args->n);
+}
+
+static void rsd_convolve1d_avx512_invoke(void *call) {
+    RsdConvolve1dCall *args = (RsdConvolve1dCall *)call;
+    rsd_convolve1d_avx512(args->a, args->na, args->b, args->nb, args->out);
+}
+
+static const RsdKernelDef rsd_avx512_kernels[] = {
+    {RSD_OP_COUNT_NONZERO, RSD_SIG_RAW_COUNT, rsd_count_nonzero_avx512_invoke},
+    {RSD_OP_CONVOLVE1D, RSD_SIG_F64_CONVOLVE, rsd_convolve1d_avx512_invoke},
+    RSD_KERNEL_DEF_END
+};
+
 void rsd_register_avx512(RsdDispatchBuilder *builder) {
-    rsd_register_count_nonzero(builder, rsd_count_nonzero_avx512);
-    rsd_register_convolve1d(builder, rsd_convolve1d_avx512);
+    rsd_register_kernel_table(builder, rsd_avx512_kernels);
 }